Restore names of stable states. For this replaced RegulatoryNode with NodeInfo in the initial state maps, may have some sides effects...

+++ src/main/java/org/ginsim/servicegui/tool/stateinregulatorygraph/StateInRegGraphFrame.java
@@ -275,7 +275,7 @@ class StableState extends TabComponantProvidingAState {
 	}
 
 	public byte[] getState() {
-		int[] state = tableModel.getState(table.getSelectedRow());
+		byte[] state = tableModel.getState(table.getSelectedRow());
 		byte[] ret = new byte[state.length];
 		for (int i=0 ; i<state.length ; i++) {
 			ret[i] = (byte)state[i];
